requirements.txt
discord.py>=2.3
python-dotenv>=1.0
deep-translator>=1.11
